代码调整

+++ web/cloud/src/views/NetworkView.vue
@@ -6,7 +6,7 @@
 				<NavViewVue current="Network" />
 			</el-aside>
 			<el-main>
-				<el-card class="box-card" v-if="this.show_type === 0">
+				<el-card class="box-card" v-show="this.show_type === 0">
 					<el-row slot="header" class="clearfix" style="height: 20px">
 						<el-button style="float: left; padding: 3px 0" type="text" @click="show_create_network">创建网络</el-button>
 					</el-row>
@@ -27,7 +27,7 @@
 							</el-table-column>
 							<el-table-column label="操作" min-width="380">
 								<template #default="scope">
-									<el-button @click="show_network_info(scope.row)" type="" size="mini">网络详情</el-button>
+									<el-button @click="show_network_info_click(scope.row)" type="" size="mini">网络详情</el-button>
 									<el-button @click="register_network(scope.row)" type="success" size="mini">重新注册</el-button>
 									<el-button @click="pasue_network(scope.row)" type="warning" size="mini" v-if="scope.row.status !== 3">开始维护</el-button>
 									<el-button @click="destroy_network(scope.row)" type="danger" size="mini">销毁网络</el-button>
@@ -36,7 +36,7 @@
 						</el-table>
 					</el-row>
 				</el-card>
-				<el-card class="box-card" v-if="this.show_type === 1">
+				<el-card class="box-card" v-show="this.show_type === 1">
 					<el-row slot="header">
 						<el-page-header @back="show_network_list" content="网络详情"></el-page-header>
 					</el-row>
@@ -64,9 +64,32 @@
 								<el-tag :type="show_network.status === 2 ? 'success' : 'danger'">{{ get_network_status(show_network) }}</el-tag>
 							</el-descriptions-item>
 						</el-descriptions>
+
+						<br />
+						<el-tabs type="border-card">
+							<el-tab-pane label="系统组件">
+								<el-table :data="system_guests" style="width: 100%">
+									<el-table-column label="ID" prop="guestId" width="80" />
+									<el-table-column label="实例名" prop="name" width="200" />
+									<el-table-column label="标签" prop="description" width="200" />
+									<el-table-column label="IP地址" prop="guestIp" width="150" />
+
+									<el-table-column label="状态" prop="status" width="100">
+										<template #default="scope">
+											<el-tag :type="scope.row.status === 2 ? 'success' : 'danger'">{{ get_guest_status(scope.row) }}</el-tag>
+										</template>
+									</el-table-column>
+									<el-table-column label="操作">
+										<template #default="scope">
+											<el-link type="primary" :href="`/#/Guest?id=${scope.row.guestId}`">详情</el-link>
+										</template>
+									</el-table-column>
+								</el-table>
+							</el-tab-pane>
+						</el-tabs>
 					</el-row>
 				</el-card>
-				<el-card class="box-card" v-if="this.show_type === 2">
+				<el-card class="box-card" v-show="this.show_type === 2">
 					<el-row slot="header">
 						<el-page-header @back="show_network_list()" content="创建网络" style="color: #409eff"></el-page-header>
 					</el-row>
@@ -142,7 +165,7 @@
 	</div>
 </template>
 <script>
-import { getNetworkList, getNetworkInfo, pauseNetwork, registerNetwork, destroyNetwork, createNetwork } from '@/api/api'
+import { getNetworkList, getNetworkInfo, pauseNetwork, registerNetwork, destroyNetwork, createNetwork, getSystemGuestList, getGuestInfo } from '@/api/api'
 import Notify from '@/api/notify'
 import NavViewVue from './NavView.vue'
 import HeadViewVue from './HeadView.vue'
@@ -155,8 +178,12 @@ export default {
 	data() {
 		return {
 			data_loading: false,
-			show_type: 0,
+
+			current_loading: false,
+			current_network_id: 0,
+			show_type: -1,
 			show_network: {},
+			system_guests: [],
 			create_network: {
 				name: '',
 				startIp: '',
@@ -172,7 +199,31 @@ export default {
 			networks: []
 		}
 	},
-	created() {
+	mounted() {
+		this.current_network_id = this.$route.query.id
+		if (this.current_network_id) {
+			this.show_type == 2
+			this.current_loading = true
+			getNetworkInfo({ networkId: this.current_network_id })
+				.then((res) => {
+					if (res.code === 0) {
+						this.show_network_info_click(res.data)
+					} else {
+						this.$alert(`获取网络信息失败:${res.message}`, '提示', {
+							dangerouslyUseHTMLString: true,
+							confirmButtonText: '返回',
+							type: 'error'
+						}).then(() => {
+							this.show_type = 0
+						})
+					}
+				})
+				.finally(() => {
+					this.current_loading = false
+				})
+		} else {
+			this.show_type = 0
+		}
 		this.init_view()
 		this.init_notify()
 	},
@@ -189,6 +240,14 @@ export default {
 				.finally(() => {
 					this.data_loading = false
 				})
+		},
+		async load_system_guest(network) {
+			this.system_guests = []
+			await getSystemGuestList({ networkId: network.networkId }).then((res) => {
+				if (res.code === 0) {
+					this.system_guests = res.data
+				}
+			})
 		},
 		get_network_status(network) {
 			switch (network.status) {
@@ -214,6 +273,26 @@ export default {
 					return 'Vlan网络'
 				default:
 					return `未知类型[${network.type}]`
+			}
+		},
+		get_guest_status(guest) {
+			switch (guest.status) {
+				case 0:
+					return '正在创建'
+				case 1:
+					return '正在启动'
+				case 2:
+					return '正在运行'
+				case 3:
+					return '正在停止'
+				case 4:
+					return '已停止'
+				case 5:
+					return '重启中'
+				case 6:
+					return '虚拟机错误'
+				default:
+					return `未知状态[${guest.status}]`
 			}
 		},
 		get_parent_network(network) {
@@ -242,6 +321,26 @@ export default {
 							this.networks.splice(findIndex, 1)
 						}
 					}
+					this.$forceUpdate()
+				})
+			} else if (notify.type === 1) {
+				getGuestInfo({ guestId: notify.id }).then((res) => {
+					if (res.code == 0) {
+						if (res.data.type === 0 && res.data.networkId === this.show_network.networkId) {
+							let findIndex = this.system_guests.findIndex((item) => item.guestId === res.data.guestId)
+							if (findIndex >= 0) {
+								this.$set(this.system_guests, findIndex, res.data)
+							} else {
+								this.system_guests.push(res.data)
+							}
+						}
+						this.$forceUpdate()
+					} else if (res.code == 2000001) {
+						let findIndex = this.system_guests.findIndex((v) => v.guestId === notify.id)
+						if (findIndex >= 0) {
+							this.system_guests.splice(findIndex, 1)
+						}
+					}
 				})
 			}
 		},
@@ -254,8 +353,9 @@ export default {
 			}
 			this.show_type = 2
 		},
-		show_network_info(network) {
+		show_network_info_click(network) {
 			this.show_network = network
+			this.load_system_guest(network)
 			this.show_type = 1
 		},
 		create_network_click() {
@@ -276,40 +376,64 @@ export default {
 			})
 		},
 		pasue_network(network) {
-			pauseNetwork({ networkId: network.networkId }).then((res) => {
-				if (res.code === 0) {
-					this.update_network_info(res.data)
-				} else {
-					this.$notify.error({
-						title: '错误',
-						message: `暂停网络失败:${res.message}`
-					})
-				}
+			this.$confirm('维护网络, 是否继续?', '提示', {
+				confirmButtonText: '确定',
+				cancelButtonText: '取消',
+				type: 'warning'
 			})
+				.then(() => {
+					pauseNetwork({ networkId: network.networkId }).then((res) => {
+						if (res.code === 0) {
+							this.update_network_info(res.data)
+						} else {
+							this.$notify.error({
+								title: '错误',
+								message: `暂停网络失败:${res.message}`
+							})
+						}
+					})
+				})
+				.catch(() => {})
 		},
 		register_network(network) {
-			registerNetwork({ networkId: network.networkId }).then((res) => {
-				if (res.code === 0) {
-					this.update_network_info(res.data)
-				} else {
-					this.$notify.error({
-						title: '错误',
-						message: `注册网络失败:${res.message}`
-					})
-				}
+			this.$confirm('重新注册网络, 是否继续?', '提示', {
+				confirmButtonText: '确定',
+				cancelButtonText: '取消',
+				type: 'warning'
 			})
+				.then(() => {
+					registerNetwork({ networkId: network.networkId }).then((res) => {
+						if (res.code === 0) {
+							this.update_network_info(res.data)
+						} else {
+							this.$notify.error({
+								title: '错误',
+								message: `注册网络失败:${res.message}`
+							})
+						}
+					})
+				})
+				.catch(() => {})
 		},
 		destroy_network(network) {
-			destroyNetwork({ networkId: network.networkId }).then((res) => {
-				if (res.code === 0) {
-					this.update_network_info(res.data)
-				} else {
-					this.$notify.error({
-						title: '错误',
-						message: `删除网络失败:${res.message}`
-					})
-				}
+			this.$confirm('销毁网络, 是否继续?', '提示', {
+				confirmButtonText: '确定',
+				cancelButtonText: '取消',
+				type: 'warning'
 			})
+				.then(() => {
+					destroyNetwork({ networkId: network.networkId }).then((res) => {
+						if (res.code === 0) {
+							this.update_network_info(res.data)
+						} else {
+							this.$notify.error({
+								title: '错误',
+								message: `删除网络失败:${res.message}`
+							})
+						}
+					})
+				})
+				.catch(() => {})
 		}
 	}
 }
